requirements.txt
numpy~=1.26.2
pandas~=2.1.3
scipy~=1.11.3
requests~=2.31.0
telebot~=0.0.5
matplotlib~=3.8.1
pycairo~=1.25.1
psycopg2-binary~=2.9.9